moved spinner
