+++ libs/designsystem/spinner/src/spinner.component.html
