+++ libs/designsystem/spinner/src/spinner.component.scss
