+++ libs/designsystem/spinner/src/spinner.component.ts
