+++ libs/designsystem/spinner/src/spinner.module.ts
