+++ libs/designsystem/src/lib/components/page/page.module.ts
@@ -1,9 +1,9 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { IonicModule } from '@ionic/angular';
+import { SpinnerModule } from '@kirbydesign/designsystem/spinner';
 
 import { FitHeadingModule } from '../../directives/fit-heading/fit-heading.module';
-import { SpinnerModule } from '../spinner/spinner.module';
 
 import { PageFooterComponent } from './page-footer/page-footer.component';
 import {
