+++ libs/designsystem/tsconfig.spec.json
@@ -13,7 +13,9 @@
       "@kirbydesign/designsystem/card": ["card/index.ts"],
       "@kirbydesign/designsystem/types": ["types/index.ts"],
       "@kirbydesign/designsystem/testing": ["testing/index.ts"],
-      "@kirbydesign/designsystem/flag": ["flag/index.ts"]
+      "@kirbydesign/designsystem/flag": ["flag/index.ts"],
+      "@kirbydesign/designsystem/spinner":["spinner/index.ts"],
+
     }
   },
   "files": ["src/test.ts"],
